requirements.txt
boto3~=1.14.16
PyJWT~=1.7.1
cryptography~=2.9.2